Fix bug after installing Proton-GE

+++ cleaner.sh
@@ -1,51 +1,51 @@
 #!/bin/bash
 
+set -e
+
+SCRIPT_FULL="$(readlink -f ${BASH_SOURCE[0]})"
+SCRIPT_PATH="$(dirname $SCRIPT_FULL)"
+SCRIPT_NAME="$(basename $SCRIPT_FULL)"
+
 NAME="SteamDeck Cache Cleaner"
-VERSION="1.0.1"
+VERSION="1.0.2"
 
-WORKPATH=$(readlink -f $(dirname $0))
-MEDIA=("/run/media")
-
-STEAM=$HOME"/.steam/steam"
-IDDB=$WORKPATH"/iddb"
-SHORTCUTS=$STEAM"/userdata/??*/config/shortcuts.vdf"
+IDDB="$SCRIPT_PATH/iddb"
+STEAM="$HOME/.steam/steam"
+SHORTCUTS="$STEAM/userdata/??*/config/shortcuts.vdf"
 CLEAN=("compatdata shadercache")
+MEDIA=("/run/media")
 
 function find_paths()
 {
-    PATHS+=($(readlink -f $STEAM"/steamapps"))
+    PATHS+=($(readlink -f "$STEAM/steamapps"))
 
-    for media in $MEDIA
-    do
-        for dir in $media/*
-        do
-            if [[ -d "$dir/steamapps" ]]
-            then
-                PATHS+=($(readlink -f $dir/steamapps))
+    for media in $MEDIA; do
+        for dir in $media/*; do
+            local steamapps="$dir/steamapps"
+            if [[ -d $steamapps ]]; then
+                PATHS+=($(readlink -f $steamapps))
             fi
         done
     done
 
-    echo "PATHS=("${PATHS[*]}")"
+    echo "PATHS=("${PATHS[@]}")"
 }
 
+# https://developer.valvesoftware.com/wiki/Add_Non-Steam_Game
 function map_shortcuts()
 {
     echo "Map shortcuts:"
+    local id_offset=6
+    local name_offset=19
 
-    # https://developer.valvesoftware.com/wiki/Add_Non-Steam_Game
-
-    id_offset=6
-    name_offset=19
-
-    positions=$(strings -t d $SHORTCUTS \
-        | grep appid                    \
+    local positions=$(strings -t d $SHORTCUTS \
+        | grep appid \
         | grep -oE '[0-9]+')
-    for p in $positions;
-    do
-        id=$(od -A n -l -j $(($p+$id_offset)) -N 4 $SHORTCUTS \
+
+    for p in $positions; do
+        local id=$(od -A n -l -j $(($p+$id_offset)) -N 4 $SHORTCUTS \
             | sed -e 's/^ *//g')
-        name=$(strings -t d $SHORTCUTS     \
+        local name=$(strings -t d $SHORTCUTS \
             | grep -w $(($p+$name_offset)) \
             | sed -e 's/^[ 0-9]* //g')
         sed -i "/$id/d" $IDDB
@@ -56,36 +56,32 @@ function map_shortcuts()
 function map_manifests()
 {
     echo "Map manifests:"
+    for path in $PATHS; do
+        for file in $path/appmanifest_*.acf; do
+            awk \
+            '{
+                if($1 == "\"appid\"") {
+                    $1="";
+                    id=gensub(/ *" */, "", "g", $0);
+                }
 
-    for file in $1/appmanifest_*.acf;
-    do
-        awk '{
-            if($1 == "\"appid\"") {
-                $1="";
-                id=gensub(/ *" */, "", "g", $0);
+                if($1 == "\"name\"") {
+                    $1="";
+                    name=gensub(/ *" */, "", "g", $0);
+                }
             }
 
-            if($1 == "\"name\"") {
-                $1="";
-                name=gensub(/ *" */, "", "g", $0);
-            }
-        }
-
-        END {
-            print name"\t"id;
-        }' $file | tee -a $IDDB
+            END {
+                print name"\t"id;
+            }' $file | tee -a $IDDB
+        done
     done
 }
 
 function map_ids()
 {
     map_shortcuts
-
-    for path in $PATHS;
-    do
-        map_manifests $path
-    done
-
+    map_manifests
     sort -u $IDDB -o $IDDB
 }
 
@@ -93,25 +89,20 @@ function prepare_info()
 {
     reg_exp="^[0-9]+$"
 
-    for path in $PATHS;
-    do
-        for clean in $CLEAN;
-        do
-            for i in $path/$clean/*
-            do
-                id=$(basename $i)
-                if [[ $id =~ $reg_exp ]]
-                then
-                    name=$(grep $id $IDDB | cut -f 1)
-                    size=$(du -h -d 0 $i | cut -f 1)
+    for path in $PATHS; do
+        for clean in $CLEAN; do
+            for file in $path/$clean/*; do
+                id=$(basename $file)
+                if [[ $id =~ $reg_exp ]]; then
+                    name=$(grep $'\t'$id$ $IDDB | cut -f 1)
+                    size=$(du -h -d 0 $file | cut -f 1)
                     type=${clean::6}
 
-                    if [[ -z $name ]]
-                    then
+                    if [[ -z $name ]]; then
                         name="Unknown"
                     fi
 
-                    INFO+=("1\t$name\t$id\t$size\t$type\t$i\n")
+                    INFO+=("1\t$name\t$id\t$size\t$type\t$file\n")
                 fi
             done
         done
@@ -119,13 +110,15 @@ function prepare_info()
 
     IFS=$'\n'
     INFO=($(sort <<< "${INFO[*]}"))
-    INFO=($(echo -e ${INFO[*]}))
+    INFO=($(echo -e ${INFO[@]}))
     unset IFS
+
+    echo "INFO=(${INFO[@]})"
 }
 
 function show_info()
 {
-    IFS=$'[\t|\n]'
+    IFS=$'[\t]'
     REMOVE=($(zenity \
         --title "$NAME $VERSION" \
         --width=1000     \
@@ -158,7 +151,7 @@ function show_confirm()
         --width=550  \
         --height=400 \
         --question   \
-        --text="Remove this folders?\n$list"
+        --text="Remove this folders?\n\n$list"
 
     res=$?
 
@@ -174,13 +167,10 @@ prepare_info
 
 REMOVE=()
 show_info
-if [[ $? == 0 && ${#REMOVE[@]} != 0 ]]
-then
+if [[ $? == 0 && ${#REMOVE[@]} != 0 ]]; then
     show_confirm
-    if [[ $? == 0 ]]
-    then
-        for path in $REMOVE
-        do
+    if [[ $? == 0 ]]; then
+        for path in $REMOVE; do
             rm -r $path
         done
     fi
